code_sandbox_light_bcdbeb7c_1769432687베타2.0/backend/requirements.txt
# Backend API Server Requirements
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dateutil==2.8.2
apscheduler==3.10.4

# Stock Data Collection
python-dotenv==1.0.0
requests==2.31.0
yfinance==0.2.35
pandas==2.1.4

# Caching (Production)
redis==5.0.1
